--- app/modules/intelligence/agents/agent_base.py
@@ -1,10 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class AgentBase(ABC):
-    def __init__(self, name: str):
-        self.name = name
-
-    @abstractmethod
-    async def run(self, input_data):
-        pass
