--- app/modules/intelligence/agents/agents_schema.py
@@ -1,8 +1,0 @@
-from pydantic import BaseModel
-
-
-class AgentInfo(BaseModel):
-    id: str
-    name: str
-    description: str
-    status: str
